Tuned the floaty-ness
Problem: Current issue is when mid air at certain angles, even with the addition of the extra gravity function the character does not descend at a normal gravitational rate.

Solution: tune the extra gravity function to be a function of angle ^2. The force had to be applied in the square function to overcome the acceleration coming from the forward movement vector being a function of transform. forward.
The extra gravity function now being a function and not a constant allows the extra gravity function to kick in as soon as the xRot angle become greater than zero opposed to the previous -20f.