Fixed Floaty-ness part 2 and Stuck on bug wall bug
Problem: moving the extra gravity to a function posed an unintentional bug where when xRot is approching the limit of 0 the extra gravity force was less than that of the regular gravity. Also when trying to exit the top of the wall run a player would get stuck due to the extra gravity function pushing them down.

Solution:
-Change the extra gravity function's added extra constant to be a higher value. Also change the power of the function to less than linear to compensate for this increase in constant.
-Added a trigger box that give an extra impulse force at the top of the wall to fix player getting stuck.